Update ROCm bindings and refactor GPUSortAllowed import paths
