+++ src/miopen/bindings.rs
@@ -41,8 +41,8 @@ pub const MIOPEN_NDEBUG: u32 = 0;
 pub const MIOPEN_ALLOC_BUFFERS: u32 = 0;
 pub const MIOPEN_ROCBLAS_VERSION_MAJOR: u32 = 5;
 pub const MIOPEN_ROCBLAS_VERSION_MINOR: u32 = 1;
-pub const MIOPEN_ROCBLAS_VERSION_PATCH: u32 = 1;
-pub const MIOPEN_ROCBLAS_VERSION_FLAT: u32 = 5001001;
+pub const MIOPEN_ROCBLAS_VERSION_PATCH: u32 = 0;
+pub const MIOPEN_ROCBLAS_VERSION_FLAT: u32 = 5001000;
 pub const MIOPEN_GOLDEN_DB_VERSION: u32 = 21;
 pub const MIOPEN_API_VERSION_REDUCE_TENSOR: u32 = 1;
 #[repr(C)]
